add new endpoint

+++ src/controllers/profile_settings.controller.js
@@ -61,6 +61,11 @@ class ProfileSettingsController {
     }
     async removeProfileSettingsArr(req,res,next){
         try{
+            const errors = validationResult(req);
+            if(!errors.isEmpty())
+            {
+                return next(ApiError.BadRequest('Error with create_profile-snaps function', errors))
+            }
             await ProfileSettingsService.removeProfileSettingsArr(req?.body);
             return res.json();
         }catch (e)
