+++ src/controllers/profile_snaps.controller.js
@@ -56,6 +56,11 @@ class ProfileSnapsController {
     }
     async removeProfileSnaps(req,res,next){
         try{
+            const errors = validationResult(req);
+            if(!errors.isEmpty())
+            {
+                return next(ApiError.BadRequest('Error with create_profile-snaps function', errors))
+            }
             await ProfileSnapService.removeProfileSnaps(req?.body);
             return res.json();
         }catch (e)
